requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==11.3.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
supabase==2.3.4
python-dotenv==1.0.0
